Refactored the aging table to reduce its memory footprint.
Converted the aging table to the use of a hikset, and suppressed the TTL field.
These changes save 4 bytes and a pointer per value stored.

Link the objects in the table together, the oldest being at the head of the
list to make it possible to expire them through a periodic event attached to
the table, instead of having a callout queue event for each value.
This saves 5 pointers per entry overall.

Therefore the total savings are 6 pointers and 4 bytes per value,