Let GGEP stream use the new zlib stream reset feature.
Whenever an extension was compressed, we used to create a deflater, use it and
then destroy it when the extension was done with it.  We are now keeing the
deflater around in case more extensions in the stream need deflation, only
reclaiming the deflater when the whole stream is closed.