Fixed GGEP incremental deflation.
This had been broken when the zlib deflater was refactored to add incremental
inflation as well.  The zlib_deflate_data() now has an additional argument
indicating whether we're done with the data to compress.

Also introduced zlib_deflate_all() to deflate the whole data configured by
a zlib_deflater_reset(), without destroying the zlib stream (so that it can
be reused for another deflation).