xmalloc_freelist_check(): changed last argument into flags.
This allows to separate emission of warning when inconsitencies are found
from the printing of the overall freelist status (mostly useful when called
from the shell).

Added the "-s" flag to the "memory check" command to request silent mode,
the default being now that each inconsistency is listed as found, in addition
to the summary line at the end.