Fixed key accounting in DBMW.
Not only is the new accounting now correct, it also avoids too many dbmap_sync()
calls by using more granular checks on the cached entries and keeping track
of the cached entries we know are definitely new (or definitely now missing)
entries compared to the physical image in the database.

This amount is also corrected each time we perform a full iteration on the
database, since we then have a unique opportunity to precisely know how many
keys there are in the physical map and how many we have dirty in the cache
and missing in the physical map.