Fixed assertion failure in ut_ack_pmsg_free().
The problem was that two EARs could be sent in a row: we were not remembering
that an EAR was sent to the lower layer, leading to possible two successive
sending.  When the second EAR got confirmed (physically sent), the assertion
triggered because we had already recorded a waiting event for the EAR reply
timeout.

The fix consists in remembering that we sent an EAR, not resending another one
until we get confirmation that the message was actually sent out.