External IP changes were not waiting for 3 hosts in 3 different CIDR space.
Code was supposed to wait for 3 different peers in 3 distinct CIDR ranges
to report a new IP before we update our advertised IP.  However, due to a typo,
this check was not done and we were just waiting for 3 notifications, from
whatever hosts (including the same one!).

Moreover, if the peer reporting the new address was in the same CIDR range
(/16 for IPv4, /64 for IPv6) as that address, we were ignoring the update!