Try to read the kernel RX UDP socket queue as quickly as possible.
When the RX queue is full, the kernel will start dropping new incoming UDP
datagrams, and we want to avoid that because this may cause us to lose an
important UDP reply, for instance.

Therefore, we allow read-ahead of messages from the UDP queue without
processing them in an attempt to leave enough room in the RX queue.  These
queued messages are then processed at a later time.