Let small hash tables be denser than larger ones.
The rationale is that small hash tables have their array of hashcodes fit
in the memory cacheline (1 or 2 entries are OK) and therefore it is not
critical to loop over more entries before finding the key since comparisons
are fast.

This lets small tables (up to 16 items) keep a small memory footprint.

Hash tables also start with an initial size of 2 items for a slow growing
rate since a lot of hash tables will end-up holding only a few items.

Further reduced the size of the hash key structure by 4 bytes (8 on 64-bit
machines) by recognizing that the amount of bits, log2(size), need only
6 bits of storage!