Reverted walloc() to use global (shared) zones for all block sizes.
This partially undoes what commit 0935821c73bff3a1282279837b453d56d2b2c469
introduced: we continue to let zalloc() handle private zones but the walloc()
wrapper is meant for general allocations and no longer requests private zones.

Indeed, it is far too difficult to guarantee that only the thread which
allocated a block with walloc() will be the one that wfree() it later. Adding
runtime checks would complicate the zalloc() and zfree() logic to the point
that it would probably alleviate the benefits of using thread-private zones
in the first place.

Now we retain the ability to request private zones for threads which will
do their own zalloc() and where it is easier to formally guarantee thread
safety via code proof-reading.