Allow any cond-waiting threads to be signalled.
Regardless of whether emulated semaphores or real ones are used, cond-waiting
threads can now receive signals whilst they are aslept at the cost of spurious
wakeups.

If it is known in advance that the thread will receive signals, it is best to
use emulated semaphores because this has less overhead and does not cause
spurious wakeups: we are able to unblock individual threads.