Moved the reference count of atoms to the tracking hash table.
Previously, the reference count was held in the atom header.  It is now stored
in the tracking hash table, which allows a more efficient runtime implementation
when compiling with PROTECT_ATOM since there are less mprotect() calls made.

Another benefit is less memory overhead for atoms on 64-bit machines:

- On 32-bit machines, there is no difference when MEM_ALIGNBYTES == 8, which is
  the typical case: we replace the unused atom header with an atom_info
  structure of 8 bytes.

- On 64-bit machines where MEM_ALIGNBYTES == 16, typically, we save 16 bytes
  per atom since we do not use the atom header any more in the regular case
  and we do not need the atom_info structure: the encoding of the reference
  count and the atom size is now made in the hash table value.

The price to pay on 64-bit machines is an additional hash table write each time
we add or remove a reference to an atom, a penalty we can avoid on 32-bit
machines because we update the atom_info structure directly.