Made the background task interface thread-safe, and introduced schedulers.
Tasks can now be put into schedulers. If no scheduler is specified (the NULL
pointer is given when creating a task) then the default scheduler is used.

If another scheduler is created and tasks are put in it, it will be up to the
application to call bg_sched_run() on the scheduler to iterate over the
registered tasks.

Added additional accessors on background tasks:

	bg_task_step() to get the task current step
	bg_task_name() to get the task name
	bg_task_wtime() to get the wall-clock time spent on task
	bg_task_step_name() to return symbolic name for the current step
	bg_task_exitcode() to fetch the exit code from the terminated task

Added bg_task_set_context() to change the user context of running tasks.
Added bgstatus_to_string() to display a symbolic version of the exit status.

Added new exit status BGS_CANCELLED to indicate that the task was cancelled.

Switched to using WCOPY_ARRAY() in bg_task_create() and bg_daemon_create().