Improved the Mersenne Twister performance by adding thread-local random pool.
The mtp_rand() routine uses a thread-local random pool whereas mt_rand() uses
a shared random pool.  The difference is that the former can use a lock-free
execution path, contrary to the latter.

Benchmarking shows a 40% throughput increase, and there is no lock contention
in our tests, so we see the cost of machine-wide synchronization used to
implement the atomic lock operations!