Callout queue now runs the idle callbacks if not scheduled in the last 30 secs.
Instead of letting the application worry about this, it is best to let the
queue decide for itself and keep track of the frequency at which the idle
callbacks are run.

Furthermore, the idle callbacks are now guaranteed to be run at most once
per second, so idle callbacks no longer need to check this.