Re-architected file_objects to allow permission checking on opened files.
All the files are now opened as O_RDWR internally, but a file_object_t is
created for each opening/creation with the proper access mode, and this
is checked for each operation on the file object.

This also allows more file descriptor sharing because there are no constraints
on the underlying kernel file descriptor, which can support both read and
write operations (provided the file permissions allow such operations, of
course).

All the created file_object_t also record the place in the source where the
file was opened, and leaking objects are dumped at exit time.