The callout queue now determines its running thread ID via cq_heartbeat().
As a result, it is no longer required to wait for the event thread to be
created before continuing, since it does not need to initialize anything
before being operational.

Also make sure the ``cq_debug_ptr'' variable is initialized to a valid
pointer, to be safe until cq_init() is called by the application layer.

The cq_dispatch() routine must now pay attention to avoid calling the
cq_heartbeat_trampoline() routine directly, or it would violate the new
cq_heartbeat() assertion that the callout queue is only run from a given
thread -- only applicable to applications where the callout queue runs in
a dedicated thread because the main thread can block.