VMM: optimized page cache management.
Page allocation now avoids taking the lock on the page cache line if that
cache line is empty.  This means doing potentially a "dirty read" on the
page count, but that's acceptable in practice.

At allocation time, we also pay attention as to how a cached region from a
higher cache line would split: if the resulting fragment ends up being put
in a cache line that is already full, then it is sub-optimal to break the
larger region.

Page cache coalescing done at freeing time is also considerably speeded up
by lazily looking at whether the cache line is empty, before grabbing the
corresponding lock.