socket_evt_set(): fixed too restrictive pre-condition on accepted events.
As written, the precondition was forbidding to set INPUT_EVENT_RW(X) which
is a valid setting, forcing clients to ask for only one of INPUT_EVENT_R(X)
or INPUT_EVENT_W(X) (i.e. was using exclusive or).

The new preconditions asks that one at least of INPUT_EVENT_R or INPUT_EVENT_W
be present.