Let thread_create_full() and thread_find() instantiate the main thread element.
We need to reserve STID = 0 to the main thread, and therefore it is imperative
that thread_create_full() creates the main element first, in case we have never
been called for the main thread before another thread is created (and we assume
that the calling thread is the main thread, of course).

Likewise, if thread_find() falls off to the bottom, assume we are in the
main thread.