dht_fill_random(): improved now that we have hash_list_random().
We no longer only select the most recently good node seen in a bucket but
any randomly chosen good node in the list for a given bucket.  The old
behaviour was flawed for passive-node DHTs since then we have a small amount
of buckets and we were limiting ourselves to one good node per bucket!