Protect some socket_evt_xxx() calls with a shell lock in case of async shell.
In code execution paths where the socket event could be changed concurrently,
protect these changes via a shell-specific lock to create a critical section.

We used to funnel back the call to shell_evt_writable() into the main thread,
but this was not correct as there was no insurance that the call would
not interrupt another event handling in the main thread for the same shell.