Prevent signal delivery when thread is waiting for a rwlock.
As soon as a thread is enqueued for a rwlock, it must not process any signal
even though it does not yet own the lock.  Indeed, should the signal handler
call a routine that needs that lock, we would re-enqueue ourselves, causing
a deadlock.

Therefore, the thread signal mask is set to block all signals as soon as the
thread is enqueued, and only when the lock will be registered in the thread
will we restore the previous signal mask, to make sure no signal can happen
until after we release the rwlock.