xmalloc: moved creation of the pool to the thread chunk free path.
Creating the pool once on the allocation path has a high potential of
recursion, and can even cause problems on 32-bit systems because we
have a nasty recursion in walloc() code, when it tries to allocate the
structure for callout queue event that monitors the pool.

We do not suffer from the same problem on the free path, and the pool is
really useful when we start to free chunks anyway.