QRP routing logic now uses LimeWire's logic, less strict than the old GTKG one.
We used to require that ALL words in a query match an entry in the QRP to
let the query pass the Bloom filter.

But LimeWire has always used a different logic: if there are more than 2 words,
then only 2/3rd of the words are required to match to let the query pass.

This means leaf nodes will get more queries than before, but at the same time
it lets queries pass that COULD match at the leaf level if the servent there
does not require that ALL the words in the query be present to return a hit.

In any case, all the other Gnutella servents out there are likely to follow
that LimeWire logic, so it is best to be compatible with them.