Updated G2 message tree formatting to align with the format used in G2 specs.
As soon as we reach the last child of a given depth level, we must no longer
output '|' vertical lines but rather a '.' to signal that we're past the
last child.  For instance:

	+ KHL
	|-+ CH
	| |-+ QK
	| \-+ V
	|-+ TS
	\-+ NH      <- last child at depth 1
	. |-+ V
	. \-+ HS

After the last /KHL/NH node, we switch from '|' to '.' in the first column.

This makes the tree more readable than emitting '|' all the way down, which
is what the code was doing before.