Fixed pmsg_t accounting to use pmsg_written_size().
The pmsg_size() routine which was used before was wrong when dealing with TCP
because the message could have been partially written to the stream first,
and pmsg_size() only refers to the amount of bytes that were flushed during
the last I/O.  To get the actual complete message size, pmsg_written_size()
must be used.

Over UDP, it does not make a difference because messages are atomically written,
but for consistency with TCP, the code was also changed in the UDP callbacks.