When targeted OOB-proxied query MUID conflicts, ensure leaf MUIDs are identical.
When OOB-proxying a query MUID "A" for a leaf and it results in a proxied MUID
"P" that is already known and attached to MUID "B", for the same leaf node,
make sure "A" and "B" are identical.  Otherwise, we would have "P" stand for
two different queries, which could lead to problems at the leaf level.

Also made OOB-proxied MUID collisions for different nodes a mandatory warning,
keeping warning for leaf MUID collisions protected by "oob_proxy_debug".  Raised
the bar for debugging messages so that running with "oob_proxy_debug" set to 1
does not produce any output other than warnings.