Make sure Gnutella messages are not accidentally forwarded to G2.
This could happen for a PUSH message that could conceivably get a route to a
neighbouring G2 hub that would be matching the GUID.  But the message cannot
be forwarded at such, it would require translation. This for upper layers to
catch and handle explicitly: at the lower level, we catch the problem and drop
the message.